requirements.txt
# Core libraries
opencv-python==4.8.1.78
insightface==0.7.3
onnxruntime==1.17.0
numpy<2.0,>=1.24.0
Pillow==10.0.0

# Web framework
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.4

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
pgvector==0.2.3

# Background tasks
celery==5.3.4
redis==5.0.0
flower==2.0.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1
pandas==2.1.1

# Monitoring
prometheus-client==0.17.1
structlog==23.1.0

# CLI and folder monitoring
watchdog==3.0.0
tabulate==0.9.0